test_install_utf8.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<window title="Mozilla Bug "
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" src="chrome://mochikit/content/chrome-harness.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script type="application/javascript" src="apphelper.js"/>
  <script type="application/javascript" src="jshelper.js"/>
  <script type="application/javascript" src="urlmatch.js"/>
  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id="
     target="_blank">Mozilla Bug </a>
  </body>

<script> 

steps = [setUp, verify_no_apps, get_installed_returns_nothing, install_bom, tearDown];

runAll(steps);

function setUp(next) { 
  info("in " + arguments.callee.name);
  uninstallAll(next);
}

function verify_no_apps(next)  {
  info("in " + arguments.callee.name);
  mozAppscb(navigator.mozApps.mgmt.getAll(), 
            [{ status: "== \"success\""}], 
            ok,
            next);
}

function get_installed_returns_nothing(next) {
  info("in " + arguments.callee.name);
  mozAppscb(navigator.mozApps.getInstalled(), 
            [{ status: "== \"success\""}], 
            ok, 
            next);
}

function install_bom(next)  {
  info("in " + arguments.callee.name);
  var appURL = SERVERS['manifest_with_bom'];
  popup_listener();
  var pending = navigator.mozApps.install(appURL, null);
  pending.onsuccess = function()  {
     var name = pending.result.manifest.name;
     var description = pending.result.manifest.description;
     ok(name == "TheBOM  ゲゴケ゚セニツ゚ヅヂチ", name);
     ok(description == "This App is THE BOM, yo. ヅヂチ", description);
     next();
  };
}

function tearDown() {
  info("in " + arguments.callee.name);
  uninstallAll();
}

</script> 

</window>
